Split sov-solana-offchain-auth main file into modules for readability
